Changed name to NodeType enumeration to include SPBN Strict nodes that are completely non-parametric.

+++ pgmpy/factors/continuous/LinearGaussianCPD.py
@@ -74,7 +74,7 @@ class LinearGaussianCPD(BaseFactor):
         self.evidence = evidence
         self.variables = [variable] + evidence
 
-        self.beta = np.asarray(beta)
+        self.beta = np.asarray(beta, dtype=np.float64)
         self.variance = variance
 
         super(LinearGaussianCPD, self).__init__()
